Face-Recognition-Based-Attendance/requirements_updated.txt
opencv-python
tensorflow==2.10.0
pymysql
pandas
pillow
gtts
apscheduler
numpy<2
scikit-learn
keras-facenet
